Split sqlite storage
